+++ openfl/experimental/interface/__init__.py
@@ -3,7 +3,7 @@
 
 """openfl.experimental.interface package."""
 
-from .fl_spec import FLSpec
+from .fl_spec import FLSpec, final_attributes
 from .participants import Aggregator, Collaborator
 
-__all__ = ["FLSpec", "Aggregator", "Collaborator"]
+__all__ = ["FLSpec", "final_attributes", "Aggregator", "Collaborator"]
