+++ openfl/experimental/runtime/__init__.py
@@ -1,8 +1,10 @@
 # Copyright (C) 2020-2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-"""openfl.experimental.runtime package."""
+""" openfl.experimental.runtime package Runtime class."""
 
-from .runtime import LocalRuntime, Runtime
+from .runtime import Runtime
+from .local_runtime import LocalRuntime
+
 
 __all__ = ["LocalRuntime", "Runtime"]
